fetching data from open api

+++ business/wallet/domain.go
@@ -4,9 +4,9 @@ import "context"
 
 type Domain struct {
 	Id_wallet   int
-	MoneySaved  int
-	MoneyTarget int
-	IndieWallet int
+	MoneySaved  float64
+	MoneyTarget float64
+	IndieWallet float64
 }
 
 type UseCase interface {
